net_gen: move handling of static-connection ports to Engine
So far, when connections where added or removed from a port whose
model is marked as 'static' (i.e. where changing the connections
requires reconfiguration), the handling of the reconfiguration was
left to the runtime connection management. This is pretty stupid
since the deployer can already determine that reconfiguration is
needed, and "spreads" such handling in different parts of the stack.

This commit moves this handling in Engine. The runtime connection
management will still, on connection removal, trigger the deployer.