make the state_reader disconnection a task-local error
The state reader can be disconnected for a lot of different reasons,
the main one being the shutdown / crash of a component. Since it
affects only the component itself and not the other components, it
should be treated as a task-local error. The task now emits :aborted
if it happens.