test: fix many problems with task context and composition stubbing
(1) we were running the deployer for each element of a composition
instead of building a clean IR and then deploying that. Some parts
of the composition would be non-deployable (as e.g. if their arguments
were referring to the other parts of the composition)

(2) tasks that had dynamic ports were stubbed into tasks of their
concrete model, which meant that they could not be started. The fix
here is to allow defining deployments for specialized models, and
let the ruby task process manager create the full task interface. Note
that it is harmless in normal operations as one would never deploy such
a task willingly (or, if he does, then he/she would really have tried very
hard to do something weird)

(3) CompositionChild#to_instance_requirements should really return an
InstanceRequirement object (instead of a CompositionChild object, which
is a subclass of InstanceRequirement). The reason for this is that
CompositionChild is interpreted differently during deployment than IR
(Composition.instanciate interprets a CompositionChild instance as a
pointer to another child in the composition). This commit fixes this
problem with a big hack. The real fix would be to remove the subclassing
of CompositionChild from InstanceRequirements -- which really does not
make a lot of sense in the first place.