test: deprecate the class-level API
It really looked awkward. It is now replaced by e.g. a spec-style

  it { is_self_contained }

In addition, this commit moves the subject_syskit_model accessor
to the relevant toplevel test classes, and uses it as the default
model if no models is given, e.g. in is_self_contained (for profiles)
or can_configure (for tasks).