netgen: remove long-useless tracking of permanent/mission tags
This was useful in the very first versions of the engine. Nowadays
the only place where static_garbage_collect is during system network
generation, where we don't have any relationship with the running plan

So, mark root tasks as permanent during system network generation, unmark
them just after and let Roby track the permanent/mission flags when
we merge with the underlying plan.