netgen: move handling of optional dependencies to network resolution
So far, optional dependencies would be considered "not present" if,
at instanciation time, the dependency was not directly injected.

This is limited as the system network postprocessing is therefore not
allowed to actually replace non-selected optional dependencies by
implementations and have it picked up.

This commit moves the handling to #compute_system_network. This allows
to explicitely select services with tasks explicitely marked as abstract
(used as "matchers") which can then be replaced at network merge time.
E.g.

    cmp.use('optional' => Cmp.use('src' => a_device).if_already_present)

will keep the optional dependency if the composition is present, and
will remove it otherwise