distinguish between 'I want to select nothing' and 'this entry should not be inherited'
The latter is used by composition instantiation to avoid having name-based
selection leak from top to bottom. The latter was introduced as a way
to override general selections for particular children (in particular
in the context of composition optional children). The original commit
used 'nil' for both, but we actually want different behaviours.

See in particular DependencyInjection#selection_for