replace InstanceRequirements#dependency_injection_context by a DependencyInjection object
The issue this solves is that the only way to implement #merge
with a DependencyInjectionContext is to concatenate the contexts,
which causes Task#requirements#dependency_injection_context to
grow exponentially on long-living tasks (with the memory and GC times
effect one imagines).

As used, the DependencyInjectionContext is not needed (things are
only added on the resolution stack), so resolution can be done
incrementally thus avoiding the possibility of memory leak altogether.