runtime: change handling of new/removal of static connections to something that at least looks like it makes sense
Some of the cases handled there really don't make sense to me,
like not having the model of a port that already has a connection.
Assume they are bugs for now, and move on from there when we actually
have use (and test) cases for them.

This specifically now handles both input and output static ports
properly. It also tests for connections addition/removal on the
ActualDataFlow -- the test on DataFlow that was done before was
actually not triggering for source or sinks that were already
garbage collected (!).

Finally, we remove the corner case of a task that is connected
to another task that is not part of the transaction. It's outside
Syskit scope for now.