runtime: fix the longstanding issue between connection management and tasks that need connections during configuration
The current configuration management scheme was to wait for all tasks
to be setup before we connect anything. This fails for tasks that pass
through proxies, as e.g. device drivers, which require the connections
to be established *and* the proxy to be started and configured before
they can go on.

This commit relaxes the scheme by creating all *static* connections
that involve at least one stopped task. These connections are harmless
for the overall network behaviour (since one of the two peers is
stopped). Moreover, tasks are now configured only when these connections
have been created.

Then, the rest of the connections (and, in particular, the connections
that involve dynamic ports) must be created only when the task is
configured (since the configuration is what creates the ports in the
first place).