actions: improve interaction between profiles and action interfaces
The main goal here is to provide a way to "dynamically" overload
requirements in actions defined as methods, that were imported using
profiles.

The main user-visible change is that the profile-defined action methods
now return an action object (which report to #as_plan and can therefore
be used in place of tasks in all plan-building APIs).

The main underlying change is the delegation of the requirement-to-action
to the to_action_model and to_action methods which already existed on
InstanceRequirements but were not used in
InstanceModelExtension#use_profile (it had its own copy :()