test: clean up coverage handling
Use the handling from Roby, instead of reimplementing it in Syskit,
and do not enable it by default in any condition (it takes multi-GB
of memory).

Also, do not set the logger's level to DEBUG by default in Test::Self,
but handle it instead in test/suite.rb. Set it to either DEBUG or
FATAL+1 randomly if coverage is not enabled, and to DEBUG if coverage
is enabled.