configure com busses to handle all their devices (even unattached ones) by default
Reconfiguring a com bus is really something that should be avoided at
all costs. The current policy would cause a reconfiguration each time
a device on the bus would be activated for the first time, which would
disrupt the already running devices.

The default policy, after this commit, is to create all dynamic
services on the bus that are required for all declared devices. One can
get the earlier behaviour by adding lazy_dispatch: true to the com bus
declaration.