runtime: fix disconnection of ports if the port tasks are still connected through other ports
This really is an artifact of the way connections are represented.
The problem this commit fixes is that the removal of a port
connection is really a edge update, not an edge removal, in the
DataFlow graph. Special care was given to add new connections, but
never to remove them.