runtime: fix corner case in connection management
ConnectionManagement#update would store delayed removed connections
in DataFlow.pending_changes, which in turn would be in the following
cycles. So far, this object was an Array, which can be fed back to
ConnectionManagement#apply_connection_changes.

However, if in the meantime some dangling connections appear, the
corresponding code snippet is doing

   Flows::DataFlow.pending_changes[2].merge!(dangling)

which expects the removed set to be a hash.

Fix by making #partition_early_late return a Hash for the late parts,
so that the outside of #partition_early_late only deals with hashes