fix Models::Composition#specialize
An unfortunate API clash makes #specialize be both the API to create
composition specializations and the one (inherited from Component)
that allows to create private specialization models (if called without
any arguments)

The latter was setting name to the argument, which is the specialization
name. It was failing if no name was provided, because it was setting
it to the option hash.