runtime: do not warn about failures during disconnection if we already know the deployments are terminating
There is a (harmless) race between terminating deployments and
disconnection of the tasks these deployments support.

In some conditions (which mostly happen during tests), the disconnection
code will try to disconnect ports of tasks of a deployment that is being
destroyed. This led so far to a bunch of warnings.

Ignore the warnings if we know the deployments are quitting (since we
know a good reason why the call could fail)