net_gen: cleanup the overall Engine call structure
The call structure was heavily dependent on storing some info
in the object, by calling e.g. #prepare. But not only. It was
therefore pretty fragile.

Replace the options hash, formerly stored in #options, by keyword
arguments that are passed along as necessary. Moreover, the
deployed_models is not computed mandatorily anymore, since it
is only needed if a TaskAllocationFailed exception is raised
(it is computed within the exception's initialize). The
task_context_deployment_candidates is now computed at the top
of #compute_deployed_network, since it is only used by this
method and the methods it calls. If directly calling one of
these sub-method, #update_task_context_deployment_candidates
should be called explicitely.