runtime: properly update connection handling when a composition child is being removed
When a composition child is removed, the corresponding forwarding need
to be removed as well. This patch does this, and ensures that the
composition will not fail in #dataflow_change_handler when the child's
exported port gets disconnected.

This was found during plan manipulation. Given that there is no
definite ordering between the graphs when removing an object
from the plan (or merging it). So, we can have the dataflow graph
removed after the dependency graph, which leads to a NoSuchChild
error in dataflow_handler