test: implement another way to stub networks
The current way to stub networks was to analyze the requirements
to stub each (inferred) element in the final network.

This inverts the method, by generating a network and modifying
it to make it configurable/startable.

Let's keep both for now, I'll probably try to switch to the new
method in the long run.