runtime: fix corner case in removal of dangling tasks
The connection notification (through #modified_tasks) is asymmetric,
since it relies on the source tasks being added when they are
disconnected from DataFlow. This fails, however, if the source task
is finalized before it gets disconnected.

There was already a special handling of dangling tasks, but they
would handle only the ones where both source and sinks are finalized.
Handle the case where the source is finalized and the sink not.