runtime: dangling connection support can now handle dead deployments, remove specific support
This fixes a race between a deployment shutdown and the removing/removed
handlers (which are triggered when the orocos tasks get disconnected).
In the current implementation, the deployment shutdown was removing
the orocos tasks from the ActualDataFlow graph, leading to the
handlers not being called (and corresponding reconfiguration not being
done in the case of static ports). Specific support in Deployment is
now removed, and dead deployments go through the normal code path in
Runtime::ConnectionManagement.