test: bring the new-style deploy-and-stub method closer to the old one
Allow for the creation of a task model on the fly
  syskit_stub_and_deploy('ModelName') do
     output_port 'out', '/double'
  end

syskit_stub_network removes the original network, the
way a deployer would do. Moreover, both network generation
and stubbing are done in transaction context, to limit their
effect to the tasks given by the caller.

In addition, the old syskit_stub methods got a _requirements suffix
(and syskit_stub is deprecated). This makes the difference clearer.