runtime: build the orocos_task->TaskContext mapping once in ConnectionManagement#initialize
This is required by the removal of singleton classes from the list
of registered models. However, this is also probably a much better
solution than the current implementation of
find_setup_syskit_task_context_from_orocos_task, as the current
implementation would linearly go through all registered models
(which, when they include the singleton classes, can be pretty huge),
it now goes through all the instanciated tasks (which is on average
smaller).