app: don't inherit OpenStruct in RobyApp::Configuration
It is really error-prone (more than once we have pieces of code that
pick the wrong flag). This break the functionality of storing component
configuration in the conf tree, but so be it: (1) we're preparing a major
release, (2) breakage is not silent, (3) the functionality is retained
(but under a 'orocos' field) and (3) I've never seen this functionality
being used in the field.