in dataflow methods, do not attempt to add an edge if the port mappings are empty
This makes it possible to call these methods with empty port mappings,
without triggering the sanity checks in ConnectionGraph#add_edge

The check is there to avoid having "empty" relations between tasks
in the dataflow graph. Something that would possibly be harmless,
but have performance impact on most of the connection-related
processing. The general design of ConnectionGraph is to try to avoid
empty relations in general.

I rather added the bypass in the DataFlow convenience methods to avoid
having

   source.add_sink(sink, mappings)

NOT actually add an edge. Would be very awkward.

However, it is IMO fine to have

   source.connect_ports(sink, mappings)
   !source.connected_to?(sink) # mappings was empty