app: start the process server on an ephemeral port
I'm trying to modify all Roby-started "support" servers to work
this way. This removes the need for configuration (only one port
has to be specified), and more easily allows starting multiple
Roby instances on the same machine.

Note that in this case, it's fairly easy as nothing else than Roby itself
needs access to the server.