remove blocking calls in the state handling logic of TaskContext
The handle resolution is now done once and for all by the deployment,
which also creates and keeps the state readers (since we make sure
that a reader is only used by one TaskContext at a time using the
should_configure_after constraints).

In case a task context that is not using a state port is encountered,
Syskit now uses a RemoteStateGetter adaptor that polls #rtt_state and
provides a compatible API. Note that this will be used later for
consistency checks on the state reported by the state reader.