explicitely pass the 'distance' between two ports to the underlying #connect_to call
This avoids having orocos.rb 'guessing' it behind the scenes, which
was broken recently because Syskit sets #process to nil to remove
unwanted orocos.rb.... :(

It also adds a separation between the process server name and whether
it is considered to be local or in-process. This allows to have multiple
process servers considered 'local' (as e.g. the ruby tasks).