runtime: avoid concurrent configuration of the same orocos task when one got garbaged
With the new async behaviour, it is possible that a task that is being
configured gets replaced by a new network. In this case, the engine
adds a SyskitConfigurationPrecedence relation to synchronize the
configurations.

However, the old task is also reap for garbage collection. This means
that Roby marks it as garbage and *removes all its external relations*.
The task is kept (through the Task#can_finalize?) mechanism, which allows
for state tracking on the Syskit side.

Because of the relation removal, the relation cannot be used in this very
particular instance for synchronization and we need to add an extra
mechanism. The relation still needs to stay, though, because it ensures
that we can always find which is the "last" task (i.e. the one that really
should be configured).

Note that all of this happens because Syskit bypasses Roby's normal
state tracking mechanisms. In principle, configuration should happen
during the component's start event. Syskit does it differently in order
to pre-configure tasks before they are actually needed, but it turns
out to be a fairly suboptimal choice, i.e. a case of premature
optimization.

In fine, I believe that the configuration-before-need is a good thing,
but that it should be done globally in a special handler, allowing
configuration-in-progress to be hijacked by newly started tasks. It would
simplify the whole logic tremendously (no need for all these flags), and
would remove the need to workaround bad interactions with Roby's normal
state tracking (such as this commit)