actions: fix the dance around DI injection and rebinding in the profile/actions interaction
How to say this ... huge mess.

There was a dance between
 - "plain" Definition objects (as defined with Profile#define)
 - "injected" InstanceRequirements objects, as returned by
   Profile#resolved_definition
 - representation of actions

This dance happened because of issues related with rebinding in action
interfaces (returning the "right" instance requirement object when
action interfaces get subclassed). The current "solution" was a pile
of hacks. For instance, #resolve_definition_on was introduced to
re-resolve an action's #requirements object at #plan_pattern time.

This commit fixes (I hope all ...) these hacks to have a clearer
interaction:
 - the requirements in Models::Action is now authoritative. DI injection
   happens at the point where the action model is created.
 - during rebinding, Models::Action looks for an equivalent action/
   definition in the target interface. If not found, it simply returns
   a copy of itself.
 - template computation happens in the action object, removing the
   need for an interface-wide cache.