track configuring and configured information in Deployment
This is the obvious place where this information could be stored,
given that the deployment needs to be alive for the information to
actually be up-to-date.

This avoids all the problems related to global state tracking (as
e.g. having to make sure that the information is deleted in all code
paths), and ensures that in case of bugs related to this tracking,
it is useful to kill the deployment to attempt working around the issue.