test: help developers handle actions with required arguments in #assert_can_*
These cannot be automatically tested. The developer must call
assert_can_* with the action and the arguments set explicitely.

This change allows to exclude some actions in the "generic" assert_can_*
call. In addition, it will gather all actions like this and flunk at
the very end with the whole list, helping the developer exclude
the relevant actions.