actions: change override rules to allow for rebinding when subclassing
The override rules were changed to NOT allow overrides. However, this
fails when doing

   class Subclass < Parent
      use_profile RefinedProfile
   end

since the actions and definitions from the parent are made available
at the subclassing point. This was actually not the case for definitions
earlier, as InterfaceModelExtension#profile was not importing the
parent's profile - something it should obviously do.

Note that this is going back to the pre-syskit2 override rules.