explicitely separate the various roles of #resolve
`#resolve` and `try_resolve` are the methods that allow to map
a task instance to a model (e.g. a data service), verifying the
task instance's type in the meantime.

However, it started having different roles. For instance
CompositionChild#resolve would resolve the child of a composition
given its parent. Or if CompositionChild#composition_model was a
CompositionChild itself, recursively go to the root and resolve the
whole chain.

Which would work OK for some things, but break in other cases, as
usual when a method has a bunch of slightly different roles. Such
case was the runtime error when resolving a grandchild in a coordination
object, e.g.

    task_script do
       a_child.b_child.out_port
    end

This commits splits #try_resolve, #try_resolve_child and
try_resolve_child_recursive for the three cases, fixing the
overall mess.