allow overriding a port period at the IR level
This allows to call #add_port_period to set a port's period, which
is then used as-is during dataflow dynamics computation. The value
overrides any value that could have been computed during computation.

It is typically be done in a profile, in cases where propagating
from the devices is too limited.