test: fix error reporting in case the errors cause the removal of the guards to raise
This is really an artifact of how the tests are built. Bugs that leave
the plan in an error state may cause the #remove_free_event to raise,
which "hides" the real error, replacing it by the error raised
by #remove_free_event's call to #unreachable (that, itself, calls
the event loop).

It's difficult to filter out the errors, so ignore the one from the
event removal - there's really nothing that should be able to cause
an error there anyways.