do not do any state-related reading in a task that has an unmet configuration precedence constraint
We don't need to. Moreover the configuration precedence constraint is
used to sequence the stop of a task with the configuration of another.

In the latter case, not properly sequencing system setup lead
the new task to read the stopped state before the running task, and
thus having the runing task not stop and the transition to never
happen.