fix deregistration of backward-compatible constants in some cases
This follows up a patch in metaruby regarding the deregistration of
constants.

The underlying problem is that deregistration must happen in two
places, in #clear_model for the receiver as well as in #clear_submodels
for children. There was no easy way to overload either one or the other
and get a consistent behaviour.

MetaRuby now introduces #clear_registration_as_constant for this purpose,
and the removal of the backward-compatible constant has been moved
there.