refactor #update_orogen_state to avoid unnecessary calls to #connected?
The only time we may get a NO_DATA is if the reader is disconnected.
So, since we have to read, read first and only check for #connected?
if the read returns NO_DATA.

This looks like a micro-optimization, but this method is really on
the very-very-hot path and these calls actually have an effect.