interface: add a x-roby-error fields with the error type
This allows the other side to recognize errors and possibly report them
differently. The name is lowercase'd because either the client API we
use in the vscode-rock extension, or grape, seems to lowercase the
headers.  This makes sure we're immune to variability in the libraries
(unless some of them uppercases, but that sounds unlikely)