ide: fix interaction between reloading models and runtime view
This is what was causing the "accessing disposed task context" error.
Reloading the models calls Orocos.clear, which in turn de-initializes
Orocos. It wasn't getting initialized again afterwards, and broke
everything.

There are other side effects - namely, that the name services are
also deinitialized. This creates new name services to avoid this.