net_gen: fix port sampling in slave activities
When the underlying activity is periodic, the dataflow dynamics
"samples" a port's input triggers at the component's fixed
period before they are propagated to the output port. In the
case of a slave, it was using the slave's activity instead of
the master's. This commit makes sure the master's period is used.