net_gen: fix corner case in flow dyn. with master/slave tasks
When the master had open triggers after #initial_information, the
code would pass twice through the slave initial information code,
which would fail.

This was caused by the current guard, which was actually using whether
a task had final information to decide whether it was explored or not.
Move to using the master/slave relationship instead.