models: clean mixup between #resolve and #bind
'bind' and 'resolve' are somewhat being used interchangeably
in the codebase. This attemps to clean up the usage of both,
with the following semantics.

'bind' "binds" a model and an already existing instance, returing
the part of the instance that is representative of the model.

   srv_m = DataService.new_submodel
   task_m = TaskContext.new_submodel
   task_m.provides srv_m, as: 'srv'
   srv_m.bind(task_m) # => task_m.srv_m

'resolve' will more generally resolve relationships between objects.
For instance, CompositionChild#resolve_and_bind_child will 'resolve'
the child from a root composition and then bind it to the composition
child model.

There is an inconsistency in the behavior of InstanceRequirements#bind,
though, which always returns its argument (if it is valid), **not**
providing a mapping between the model services and ports to the
argument's actual ports.

However, since #bind is used for the return value of
InstanceRequirements#instanciate, changing this would have far-reaching
consequences that I can't deal with right now

I elected to keep the old behavior until I have the time to dig into it