fix a corner case in handling of composition-child relationships within dataflow triggers
One of the dataflow triggers assumed that children of a composition
only have forwarding (in this case, input-to-input) connections to
the composition's own ports.

This is usually true, but dependency injection and/or network merging
can cause a child to be both part of a composition, and a user of the
composition's output, which caused a bug within the dataflow handler.