app: load orocos/async if Conf.ui is set
This is a new flag set by the --ui argument to `syskit test`

The goal is to allow for 'interactive' UI tests, where the UIs are
displayed and we use the user as the 'evaluator' of the result.
Worst than pure tests, but better than "manual" tests.

The need to have it handled in Syskit is that there are requirements
on when orocos/async can be loaded (namely, it has to be loaded before
Orocos.initialize is called)