pass connection policies as a concrete graph from net gen to runtime
This is actually a fix for a much worse bug that the title would let
one think. Since f1672d01fce6d16f5dc1eda3093cf2ae326ce03e
(merged 2019/01), Syskit would not apply *any* of the connection
policies computed during net gen to the runtime plan. Meaning, all
connections except logging would be using data.

Yep.

I found this out while trying to debug a CAN-based driver that
had to deal with bursts of 10 packets, but only seeing 1/3 of them
coming in, regardless of the computed buffer size.

The bug was created by having one line of the commit commented out
(because it wasn't implemented :() and no integration tests related
to this particular issue. This commit fixes both.

The general idea of the change was a good one, though. The problem
it tried to solve was that policy computation would change **one**
element in DataFlow, which would lose information (what are the
requirements exactly ?) and makes it harder to reason on the result.

Since policies are really computed for concrete connections, it's now
computed as a hash over the task/ports and passed on to the runtime
subsystem along with the connection graph. This is possible
since we're computing the whole graph each time. Entries that are
not present use the "old" method based on the DataFlow graph, to
allow for connections manually added to the DataFlow graph after
deployment.