fix multiple bugs in flow dynamics computation related to busses
Bursts were bizarrely handled, fed as a trigger of size 1 and period
"dev.period * dev.burst".  I can't find any justification for this in
the history, so reset it to a period of zero (meaning "count one
burst").

The code iterates on ports from the device service, and was adding the
port name as declared in the service to handled_ports.  This meant that
the wrong port name was passed to done_port_info later in the method and
could lead to a failure in the end.

Last, but not least, the code was iterating on inputs, not outputs which
makes no sense given that a device periods characterizes the data it
produces.