net_gen: modify the deployment group propagation logic to something usable
Deployment groups were introduced to refactor the way deployments
were declared and managed.

Pre-groups, all deployments were declared at the level of the global
configuration, and chosen using `prefer_deployed_tasks`. This showed
serious limitations (apart from the ugliness of the globality of it
...). The two main usability issues were:

1. deployments are defined in the robot config and used in profiles. The
   link between the two is implicit. Reusing a deployed profile turned
   out to be difficult, as was "instanciating" a profile multiple times
2. Locally "massaging" instance requirements to change how things are
   deployed without affecting the rest of the system is close to
   impossible.

Deployment group's propositions was to create an object to manage and
resolve deployments locally, and attach a group to a task's (through
InstanceRequirements). The system deployer would be asking these
objects for deployments _et voila_. A global deployment group would
be used as fallback to keep things running.

However, the current implementation would resolve a "deployment group"
that is a top-down merge of the groups present in the plan: start with
the root of the hierarchy, the groups would be merged (as in: union-ed)
until the leaves (the tasks) and then deployments would be resolved.
This made overriding (obviously) impossible and forbid gradually
moving to groups (if a deployment is available for a task globally, it
would create ambiguities when another one is defined more locally).

This changes the implementation to instead look for the first matching
deployment bottom-up, and stop there. Overriding becomes the default.

With this change (and the new #deployed_as helper), one can define an
deploy a task (e.g. a device) with

~~~ ruby
define 'camera', camera_dev.deployed_as('front_camera')
~~~

Ambiguities might still happen if there is a fork in the hierarchy tree:

- A depends on task
- B depends on task
- both A and B define different deployments for task

This could be solved by applying deployments _before_ merging the
network(s), which would also allow us to force multiple instanciations
of the same task by assigning multiple deployments. But that's for
another time.