test: fix syskit_deploy to allow specific requirements to override the test-level deployment group
This changes how syskit_deploy pushes the test-level deployment group.
It was before pushed on the instance requirements before deployment,
which meant that it had 100% precedence over everything else, including
what was written in the test.

This changes it so that the global group + test group is pushed as
a default group to the deployer instead. Local groups can then be set
if more fine-grained setup is needed