test: let the test writer(s) decide whether an abstract test should run live or not
This boils down to the "subject model" concept in Roby/Syskit
tests. In task context tests, Syskit tries to auto-define a
deployment for the test subject, e.g.

~~~
describe OroGen.iodrivers_base.ProxyTask do
end
~~~

This definition is not possible when in live mode for abstract
models (e.g. iodrivers_base.Task). For this reason, the current
spec implementation in Syskit was skipping these tests automatically.

However, one can define test tasks based on these abstract models
to test the abstract task functionality. In the case of these
test components, we want the subject model to be the abstract
task but run the tests live using a test task written for this
purpose.

Since Roby allows to define whether a test suite should run live
or not, let the developer decide instead of enforcing the behavior
deep down in the test implementation.