test: add helpers to create reader and writers during tests
And disconnect them on teardown.

syskit_create_writer is necessary (as opposed to using syskit_write)
if the component behaves differently if (1) there is no connection
(e.g. iodrivers_base during configuration) or (2) if there is NoData.

syskit_create_reader allows to mock a component that would "reply"
to the task under test within a poll block, instead of having to
play with have_one_new_sample in execution expectation blocks. Not
having to do the latter makes sure that we can concentrate on the test
itself.

The "introduction" use case is the ability to unit test a component
built on top of iodrivers_base. For this component we need to interact
with the io_raw_in and io_raw_out ports to fake the actual device to
be allowed to configure and start the component.