auto-disconnect port accessors on component finalization
The accessors will now be automatically disconnected once the target
port's component gets finalized, or if the actual port component is (in
case of port exports). So far, port readers had an automatic lifetime
only when created in task scripts, and even there it could lead to
weird issues - like having a connected port but no component actually
listening to it (since the target component was stopped).

In addition, since the ports were outliving their target components,
calling `#disconnect` would fail as `#disconnect` was using the
execution engine of the port's component as resolved at disconnection
time. We now cache this value as it can't change in any case.