implement dynamic data sources
Dynamic data sources "anchor" a port dynamically in the system,
on the basis of e.g. a plan query or a composition child. They
allow to resolve and re-resolve ports as actual components come
and go, and read from them

Their goal is to provide a service-oriented view of the data
sources in the plan, for instance for the purpose of a UI or monitoring.
Using dynamic data sources, one may tag specific compositions or
components with services that are then used to 'find' the appropriate
source at any given state, in a way that's orthogonal from the data flow.