workaround ruby 2.7 "arbitrary object as keyword argument" change
In the middle of the whole "real keyword argument migration", Ruby 2.7
changed how a mix of strings and symbols in the "last hash argument" is
handled.

Namely, non-symbol keys are OK as keyword arguments now, and Ruby passes
them to the **kw splat by default, that is in

~~~
provides model, "some" => "mapping", as: "name"
~~~

The `"some" => "mapping"` hash gets into the keyword splat. Which is
rather obviously really bad for us. I've added helpers to handle
that case in Roby, and use them here to patch the problematic calls