test: fix flaky test
The create_deployment_model helper would create the tasks and the
deployments. It turns out that (unexpectedly), Syskit deployment
models would not hold a reference to the corresponding Syskit
task models (it only refers to the orogen objects), and sometimes
the syskit models would be garbage collected before they're needed.

I never saw this in the wild, but I think it is something that
should ultimately be fixed at the deployment level.