net_gen: carry options passed to Async.start/.prepare to the apply step
Some of these options (namely, compute_deployments) control whether
apply has to finalize things that were prepared in the async phase.
When set incorrectly, they can lead to failures in the deployment
phase.

Note that this does not impact the normal flow (as it uses only
default options, which are coherent). It led to failures during
tests.